msk-serverless-client/requirements.txt
confluent-kafka==2.2.0
kafka-python==2.0.2
aws-msk-iam-sasl-signer-python